# server_service/docs/swagger.yaml
definitions:
  github_com_lits-06_vcs-sms_server_service_internal_dto.CreateServerRequest:
    properties:
      ipv4:
        description: IPv4 address
        example: 192.168.1.100
        type: string
      name:
        description: Server name
        example: web-server-01
        type: string
      port:
        description: Port number
        example: 80
        maximum: 65535
        minimum: 1
        type: integer
    required:
    - ipv4
    - name
    - port
    type: object
  github_com_lits-06_vcs-sms_server_service_internal_dto.UpdateServerRequest:
    properties:
      id:
        description: Server ID
        example: uuid-string
        type: string
      ipv4:
        description: IPv4 address
        example: 192.168.1.101
        type: string
      name:
        description: Server name
        example: updated-server-name
        type: string
    type: object
  httpresponse.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
host: server.localhost
info:
  contact: {}
  description: This is the Server Service API for VCS-SMS system
  termsOfService: http://swagger.io/terms/
  title: Server Service API
  version: "1.0"
paths:
  /api/servers:
    get:
      consumes:
      - application/json
      description: Retrieve servers with optional filtering, sorting and pagination
      parameters:
      - description: Filter by server name
        in: query
        name: name
        type: string
      - description: Filter by server status (ON/OFF)
        in: query
        name: status
        type: string
      - description: Filter by IPv4 address
        in: query
        name: ipv4
        type: string
      - description: Pagination offset
        in: query
        name: from
        type: integer
      - description: Pagination limit
        in: query
        name: to
        type: integer
      - description: Sort by field (name, status, created_at, updated_at)
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of servers with total count
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "500":
          description: Failed to view server
          schema:
            $ref: '#/definitions/httpresponse.Response'
      security:
      - BearerAuth: []
      summary: View servers
      tags:
      - Server
    post:
      consumes:
      - application/json
      description: Create a new server with specified name, status, IPv4 address and
        port
      parameters:
      - description: Server data
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/github_com_lits-06_vcs-sms_server_service_internal_dto.CreateServerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Server created successfully
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "500":
          description: Failed to create server
          schema:
            $ref: '#/definitions/httpresponse.Response'
      security:
      - BearerAuth: []
      summary: Create a new server
      tags:
      - Server
    put:
      consumes:
      - application/json
      description: Update server information including name and IPv4 address
      parameters:
      - description: Server update data
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/github_com_lits-06_vcs-sms_server_service_internal_dto.UpdateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Server updated successfully
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "500":
          description: Failed to update server
          schema:
            $ref: '#/definitions/httpresponse.Response'
      security:
      - BearerAuth: []
      summary: Update a server
      tags:
      - Server
  /api/servers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a server by its ID
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Server deleted successfully
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "400":
          description: Server ID is required
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "500":
          description: Failed to delete server
          schema:
            $ref: '#/definitions/httpresponse.Response'
      security:
      - BearerAuth: []
      summary: Delete a server
      tags:
      - Server
  /api/servers/export:
    get:
      consumes:
      - application/json
      description: Export servers to an Excel file with optional filtering
      parameters:
      - description: Filter by server name
        in: query
        name: name
        type: string
      - description: Filter by server status (ON/OFF)
        in: query
        name: status
        type: string
      - description: Filter by IPv4 address
        in: query
        name: ipv4
        type: string
      - description: Pagination offset
        in: query
        name: from
        type: integer
      - description: Pagination limit
        in: query
        name: to
        type: integer
      - description: Sort by field (name, status, created_at, updated_at)
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: Excel file with servers data
          schema:
            type: file
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "500":
          description: Failed to export servers
          schema:
            $ref: '#/definitions/httpresponse.Response'
      security:
      - BearerAuth: []
      summary: Export servers to Excel
      tags:
      - Server
  /api/servers/import:
    post:
      consumes:
      - multipart/form-data
      description: Import servers from an uploaded Excel file
      parameters:
      - description: Excel file to import
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Import result with success and failure counts
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "400":
          description: File is required
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpresponse.Response'
        "500":
          description: Failed to import servers
          schema:
            $ref: '#/definitions/httpresponse.Response'
      security:
      - BearerAuth: []
      summary: Import servers from Excel
      tags:
      - Server
  /health:
    get:
      consumes:
      - application/json
      description: Check if the Server Service is running and healthy
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            $ref: '#/definitions/httpresponse.Response'
      summary: Health check
      tags:
      - Health
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"